+++ data/tasks.txt
@@ -1,6 +1,8 @@
-D|1|do w|12-08-2022
+D|1|do |12-08-2022
+D|0|sl|12-12-2022
 D|0|sle|12-12-2022
-D|0|slee|12-12-2022
 T|0|sleep
-D|0|sleep|12-12-2022
+D|0|slee|12-12-2022
+E|0|slee|12-12-2022
+T|0|sleep
 E|0|sleep|12-12-2022
