+++ src/main/java/duke/DukeEmptyDescriptionException.java
@@ -1,10 +1,10 @@
-package Duke;
+package duke;
 
 /**
  * If a command does not have any descriptions,
  * then this will be thrown.
  */
-public class DukeEmptyDescriptionException extends DukeException{
+public class DukeEmptyDescriptionException extends DukeException {
 
     /**
      * Returns a specific message depending on the command
