+++ src/main/java/duke/ToDo.java
@@ -1,4 +1,4 @@
-package Duke;
+package duke;
 
 public class ToDo extends Task {
 
@@ -8,7 +8,7 @@ public class ToDo extends Task {
 
     @Override
     public String toString() {
-        return "[T]"  + super.toString();
+        return "[T]" + super.toString();
     }
     @Override
     public String storeToString() {
