+++ text-ui-test/EXPECTED.TXT
@@ -1,4 +1,4 @@
-Hello I'm Duke, what can I do for you?
+Hello I'm duke, what can I do for you?
 Currently, you do not have any tasks.
 Added this task:
   [T][?] borrow book, to your list
